drivers: sensor: adi: ad2s1210: widen the RPM product before dividing
The velocity word is int16_t and the range from the lookup table is
int32_t, so velocity * range is evaluated in 32 bits. It overflows for
the three fastest of the eight ranges: 75000 from |velocity| 28634,
150000 from 14317, 187500 from 11454, and that last threshold is about
a third of full scale. Widen to int64_t before the divide.

+++ drivers/sensor/adi/ad2s1210/ad2s1210.c
@@ -422,9 +422,9 @@ static int ad2s1210_channel_get(const struct device *dev, enum sensor_channel ch
 	case SENSOR_CHAN_RPM: {
 		/* Get max range from lookup table */
 		int32_t range = table_velocity_range_rpm[data->resolution][data->clock];
-		/* Convert raw velocity to RPM */
-		int rpm = ((int32_t)data->velocity * range) /
-			  ((1 << (AD2S1210_MAX_RESOLUTION_BITS - 1)) - 1);
+		/* The product overflows 32 bits in the fastest ranges */
+		int rpm = (int)(((int64_t)data->velocity * range) /
+				((1 << (AD2S1210_MAX_RESOLUTION_BITS - 1)) - 1));
 
 		val->val1 = rpm;
 		val->val2 = 0;
